Rank premium

+++ Commands/ranks/del-tester.js
@@ -9,6 +9,7 @@ module.exports = {
     const member = message.mentions.members.first() || message.guild.members.cache.get(args[0])  
     const data = await tester.findOne({user: member.id})
     if(!data) return message.channel.send("No se encuentra este usuario en lista.")
+    if(!data.user) return message.channel.send(`El usuario ${member} no es tester`)
     if(data){
       await data.delete()
       return message.channel.send(`Se removio al usuario ${member} de testiadores de Torrecoin`)
